feat: refactorizacion de la vista de exámenes por partes
