+++ exam-craft-extension-frontend/src/css/Vertical.css
@@ -1,0 +1,63 @@
+.vertical-menu-container {
+  display: flex;
+  flex-direction: column; 
+  gap: 20px;             
+  width: 100%;
+  max-width: 500px;       
+  margin-bottom: 30px;
+  align-self: center; 
+}
+
+.menu-btn {
+  background-color: #A67C6D; 
+  color: white;
+  font-size: 1.3rem;
+  font-weight: bold;
+  padding: 20px 30px;     
+  border: none;
+  border-radius: 15px;    
+  cursor: pointer;
+  box-shadow: 0 4px 8px rgba(0,0,0,0.2); 
+  transition: all 0.2s ease;
+  text-transform: uppercase; 
+  letter-spacing: 1px;
+  width: 100%; 
+}
+
+.menu-btn:hover {
+  background-color: #8D5E50;
+  transform: translateY(-3px);
+  box-shadow: 0 6px 12px rgba(0,0,0,0.3);
+}
+
+.menu-btn:active {
+  transform: translateY(0); 
+  box-shadow: 0 2px 4px rgba(0,0,0,0.2);
+}
+
+.menu-btn1 {
+  background-color: #A67C6D; 
+  color: white;
+  font-size: 1.3rem;
+  font-weight: bold;
+  padding: 20px 30px;     
+  border: 3px solid #4A3728; 
+  border-radius: 15px;    
+  cursor: pointer;
+  box-shadow: 0 4px 8px rgba(0,0,0,0.2); 
+  transition: all 0.2s ease;
+  text-transform: uppercase; 
+  letter-spacing: 1px;
+  width: 100%; 
+}
+
+.menu-btn1:hover {
+  background-color: #8D5E50;
+  transform: translateY(-3px);
+  box-shadow: 0 6px 12px rgba(0,0,0,0.3);
+}
+
+.menu-btn1:active {
+  transform: translateY(0); 
+  box-shadow: 0 2px 4px rgba(0,0,0,0.2);
+}
